pyproject.toml

[project]
name = "hamstercage"
version = "0.0.1"
dependencies = [
    "pyyaml",
]
